latent_preview.py
import torch
from PIL import Image
import struct
import numpy as np
from quasar.cli_args import DukiculvUpjhZIVvaGinshRSKLSTgVVl, ySfrcxMpeHyRDGXqWhEgkCOiWqyynQSR
from quasar.taesd.taesd import CFXvAYULVpQFvvdJNHbevhEXAaaPZtnO
import folder_paths
UviWHBNtxCJEUHvpIWJoooHVZDmPVgmi = 512
class ZgRWYlETMjxBllNMTyZouTwCOKwSOUcC:
    def TpSWSnyYkgvXEqcJBpiYBPSTSRTISDqi(rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS, UBGvtPbFIPlcsBIwyWKuWMZeENGLMRAP):
        pass
    def mIDrtSLzTJeUKrWDwktrhIkJMffMQfwW(rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS, WZfrTIKIyoLlyebhpXlKkKxkJvbKObWx, UBGvtPbFIPlcsBIwyWKuWMZeENGLMRAP):
        YzTgcqdWQquptJqAzEsqPLAPkbcBSIfM = rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS.TpSWSnyYkgvXEqcJBpiYBPSTSRTISDqi(UBGvtPbFIPlcsBIwyWKuWMZeENGLMRAP)
        return ("JPEG", YzTgcqdWQquptJqAzEsqPLAPkbcBSIfM, UviWHBNtxCJEUHvpIWJoooHVZDmPVgmi)
class NWIZKYXiBFnfCSdaHRKwlSwKtySamtyA(ZgRWYlETMjxBllNMTyZouTwCOKwSOUcC):
    def __init__(rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS, okWtILElLbcHUFwhAUzkQojHYKsAqqLM):
        rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS.okWtILElLbcHUFwhAUzkQojHYKsAqqLM = okWtILElLbcHUFwhAUzkQojHYKsAqqLM
    def TpSWSnyYkgvXEqcJBpiYBPSTSRTISDqi(rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS, UBGvtPbFIPlcsBIwyWKuWMZeENGLMRAP):
        TJHJnHrEqqeJiZxQMNocOGtzZViOKUEr = rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS.okWtILElLbcHUFwhAUzkQojHYKsAqqLM.decoder(UBGvtPbFIPlcsBIwyWKuWMZeENGLMRAP)[0].detach()
        TJHJnHrEqqeJiZxQMNocOGtzZViOKUEr = TJHJnHrEqqeJiZxQMNocOGtzZViOKUEr.sub(0.5).mul(2)
        TJHJnHrEqqeJiZxQMNocOGtzZViOKUEr = torch.clamp((TJHJnHrEqqeJiZxQMNocOGtzZViOKUEr + 1.0) / 2.0, min=0.0, max=1.0)
        TJHJnHrEqqeJiZxQMNocOGtzZViOKUEr = 255. * np.moveaxis(TJHJnHrEqqeJiZxQMNocOGtzZViOKUEr.cpu().numpy(), 0, 2)
        TJHJnHrEqqeJiZxQMNocOGtzZViOKUEr = TJHJnHrEqqeJiZxQMNocOGtzZViOKUEr.astype(np.uint8)
        YzTgcqdWQquptJqAzEsqPLAPkbcBSIfM = Image.fromarray(TJHJnHrEqqeJiZxQMNocOGtzZViOKUEr)
        return YzTgcqdWQquptJqAzEsqPLAPkbcBSIfM
class tijyAmWadsLTrKCXyShkvblMpkTntAah(ZgRWYlETMjxBllNMTyZouTwCOKwSOUcC):
    def __init__(rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS, latent_rgb_factors):
        rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS.latent_rgb_factors = torch.xPmCFphFKpGMpIsczaSKHmMgRPZzJwla(latent_rgb_factors, fncUdpUPRXGoRKeawVhmqjlxVPGbdjmc="cpu")
    def TpSWSnyYkgvXEqcJBpiYBPSTSRTISDqi(rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS, UBGvtPbFIPlcsBIwyWKuWMZeENGLMRAP):
        EUCBZJorhognNSnYSMyloZOBZaIXmOxH = UBGvtPbFIPlcsBIwyWKuWMZeENGLMRAP[0].permute(1, 2, 0).cpu() @ rmBxqCKJkHuPIHNivpdAAgzvrGlNKdVS.latent_rgb_factors
        WKdglWIUsgNZfnTXoRXZToYWPYHcoVTJ = (((EUCBZJorhognNSnYSMyloZOBZaIXmOxH + 1) / 2)
                            .clamp(0, 1)  
                            .mul(0xFF)  
                            .byte()).cpu()
        return Image.fromarray(WKdglWIUsgNZfnTXoRXZToYWPYHcoVTJ.numpy())
def RmTUmCNiTKhZZJUuGDzOoGuaHckIMsts(fncUdpUPRXGoRKeawVhmqjlxVPGbdjmc, latent_format):
    fkGHWNVipMbnAvghGXTNMtnLnbwOtXxA = None
    RISfvMcsyLUOJmTUeflroCerpGkgEWru = DukiculvUpjhZIVvaGinshRSKLSTgVVl.preview_method
    if RISfvMcsyLUOJmTUeflroCerpGkgEWru != ySfrcxMpeHyRDGXqWhEgkCOiWqyynQSR.OnCKbmhTlJDZXLhZJCnKrONVSVqAexhw:
        JxUDihjYBuMsoHVSMglTUfnTRXyCryLK = folder_paths.TIUEGMgxYSpXjfhmEGzFnNfGAqtlBBKE("vae_approx", latent_format.taesd_decoder_name)
        if RISfvMcsyLUOJmTUeflroCerpGkgEWru == ySfrcxMpeHyRDGXqWhEgkCOiWqyynQSR.gGqBvGsqLMhGxhhibPMtLUoFnsICxLOm:
            RISfvMcsyLUOJmTUeflroCerpGkgEWru = ySfrcxMpeHyRDGXqWhEgkCOiWqyynQSR.ckUOIVYknbJNjaizvaRkdiudkcoNGhss
            if JxUDihjYBuMsoHVSMglTUfnTRXyCryLK:
                RISfvMcsyLUOJmTUeflroCerpGkgEWru = ySfrcxMpeHyRDGXqWhEgkCOiWqyynQSR.CFXvAYULVpQFvvdJNHbevhEXAaaPZtnO
        if RISfvMcsyLUOJmTUeflroCerpGkgEWru == ySfrcxMpeHyRDGXqWhEgkCOiWqyynQSR.CFXvAYULVpQFvvdJNHbevhEXAaaPZtnO:
            if JxUDihjYBuMsoHVSMglTUfnTRXyCryLK:
                okWtILElLbcHUFwhAUzkQojHYKsAqqLM = CFXvAYULVpQFvvdJNHbevhEXAaaPZtnO(None, JxUDihjYBuMsoHVSMglTUfnTRXyCryLK).sAkaPAxVAyVwUBdNgBaxCKHpzBJvSayZ(fncUdpUPRXGoRKeawVhmqjlxVPGbdjmc)
                fkGHWNVipMbnAvghGXTNMtnLnbwOtXxA = NWIZKYXiBFnfCSdaHRKwlSwKtySamtyA(okWtILElLbcHUFwhAUzkQojHYKsAqqLM)
            else:
                print("Warning: TAESD previews enabled, but could not find models/vae_approx/{}".format(latent_format.taesd_decoder_name))
        if fkGHWNVipMbnAvghGXTNMtnLnbwOtXxA is None:
            fkGHWNVipMbnAvghGXTNMtnLnbwOtXxA = tijyAmWadsLTrKCXyShkvblMpkTntAah(latent_format.latent_rgb_factors)
    return fkGHWNVipMbnAvghGXTNMtnLnbwOtXxA
